Added Tween class; changed update delta from float to double; added text offset animation for buttons

+++ hvn3/include/hvn3/core/UpdateEventArgs.h
@@ -6,14 +6,14 @@ namespace hvn3 {
 	class UpdateEventArgs : public System::GameStateEventArgs {
 
 	public:
-		UpdateEventArgs(float dt);
-		UpdateEventArgs(float dt, IGameManager* manager);
+		UpdateEventArgs(double dt);
+		UpdateEventArgs(double dt, IGameManager* manager);
 
 		// Returns the number of seconds elapsed since the last update.
-		float Delta() const;
+		double Delta() const;
 
 	private:
-		float _delta_time;
+		double _delta_time;
 
 	};
